started abstracting funtionality
