+++ src/components/VTabs/VTabsItems.js
@@ -5,16 +5,39 @@ export default {
 
   directives: { Touch },
 
-  inject: ['next', 'prev'],
-
   props: {
     cycle: Boolean,
     touchless: Boolean
   },
 
   methods: {
+    next (cycle) {
+      let nextIndex = this.activeIndex + 1
+
+      if (!this.tabItems[nextIndex]) {
+        if (!cycle) return
+        nextIndex = 0
+      }
+
+      this.activeIndex = nextIndex
+    },
+    prev (cycle) {
+      let prevIndex = this.activeIndex - 1
+
+      if (!this.tabItems[prevIndex]) {
+        if (!cycle) return
+        prevIndex = this.tabItems.length - 1
+      }
+
+      this.activeIndex = prevIndex
+    },
     onSwipe (action) {
       this[action](this.cycle)
+    },
+    updateContent () {
+      this.content.forEach(({ toggle }) => {
+        toggle(this.target, this.reverse, this.isBooted)
+      })
     }
   },
