+++ src/stylus/components/_tabs.styl
@@ -36,46 +36,68 @@ theme(tabs, "tabs__bar")
 
   .icon--append
     right: 4px
-  
-  // &--align-with-title,
-  // &--overflow
-  //   .tabs__wrapper
-  //     overflow: hidden
-  
-  &--overflow
+
+  &--show-arrows
     .tabs__wrapper
       margin-left: 40px
       margin-right: 40px
 
+  &--overflow      
+    .tabs__div
+      flex: 1 0 auto
+
   &--align-with-title
-    .tabs__wrapper
+    .tabs__container div:first-child,
+    .tabs__slider + .tabs__div
       margin-left: 56px
       
       @media $display-breakpoints.xs-only
-        margin-left: 66px
+        margin-left: 64px
+    
+    &.tabs__bar--show-arrows
+      .tabs__container div:first-child,
+      .tabs__slider + .tabs__div
+        margin-left: 16px
+        
+        @media $display-breakpoints.xs-only
+          margin-left: 24px
+
+  &--fixed-tabs,
+  &--icons-and-text
+    .tabs__div
+      min-width: 160px
+      
+  &--fixed-tabs,
+  &--icons-and-text,
+  &--centered
+    @media $display-breakpoints.xs-only
+      .tabs__div
+        min-width: 72px
 
   &--fixed-tabs
     .tabs__div
       min-width: 160px
 
       @media $display-breakpoints.xs-only
+        flex: 1 0 auto
         min-width: 72px
 
-  &--centered:not(.tabs__bar--is-mobile)
+  &--centered,
+  &--fixed-tabs
     .tabs__container
-      text-align: center
-  
-  // Assuming most use-cases won't break this
-  // When left as display flex on mobile,
-  // long text has issues with wrapping
+      div:first-child
+        margin-left: auto
+        
+      div:last-child
+        margin-right: auto
+        
+      .tabs__slider + .tabs__div
+        margin-left: auto
+      
   &--grow
-    @media $display-breakpoints.sm-and-up
-      .tabs__container
-        display: flex
-
-      .tabs__div
-        flex: 1 0
-        max-width: none
+    .tabs__div
+      flex: 1 0 auto
+      max-width: none
         
   &--icons-and-text
     .tabs__item
@@ -84,19 +106,19 @@ theme(tabs, "tabs__bar")
       .icon
         margin-bottom: 6px
 
+.tabs__wrapper
+  overflow: hidden
+  
 .tabs__container
+  display: flex
   list-style-type: none
   transition: transform 0.6s cubic-bezier(0.86, 0, 0.07, 1)
   white-space: nowrap
 
-.tabs__items
-  overflow: hidden
-  position: relative
-
 .tabs__div
   align-items: center
   display: inline-flex
-  flex: 0 1
+  flex: 0 1 auto
   height: inherit
   max-width: 264px
   vertical-align: middle
@@ -120,7 +142,6 @@ theme(tabs, "tabs__bar")
   transition: $primary-transition
   user-select: none
   white-space: normal
-  word-break: break-all
   
   &:not(.tabs__item--active)
     opacity: .7
@@ -131,9 +152,10 @@ theme(tabs, "tabs__bar")
   position: absolute
   transition: $primary-transition
 
+.tabs__items
+  overflow: hidden
+  position: relative
+ 
 .tabs__content
   width: 100%
   transition: transform 0.4s cubic-bezier(0.86, 0, 0.07, 1)
-  
-.tabs__wrapper
-  overflow: hidden
